use generatic type unify three executors
