+++ src/batch/src/executor/mod.rs
@@ -19,7 +19,7 @@ use risingwave_pb::batch_plan::PlanNode;
 
 use crate::executor2::{
     BoxedExecutor2, BoxedExecutor2Builder, DeleteExecutor2, ExchangeExecutor2, FilterExecutor2,
-    GenerateSeriesExecutor2Wrapper, HashAggExecutor2Builder, HashJoinExecutor2Builder,
+    GenerateSeriesExecutor2Builder, HashAggExecutor2Builder, HashJoinExecutor2Builder,
     HopWindowExecutor2, InsertExecutor2, LimitExecutor2, MergeSortExchangeExecutor2,
     NestedLoopJoinExecutor2, OrderByExecutor2, ProjectExecutor2, RowSeqScanExecutor2Builder,
     SortAggExecutor2, SortMergeJoinExecutor2, TopNExecutor2, TraceExecutor2, ValuesExecutor2,
@@ -112,7 +112,7 @@ impl<'a> ExecutorBuilder<'a> {
             NodeBody::SortMergeJoin => SortMergeJoinExecutor2,
             NodeBody::HashAgg => HashAggExecutor2Builder,
             NodeBody::MergeSortExchange => MergeSortExchangeExecutor2,
-            NodeBody::GenerateSeries => GenerateSeriesExecutor2Wrapper,
+            NodeBody::GenerateSeries => GenerateSeriesExecutor2Builder,
             NodeBody::HopWindow => HopWindowExecutor2,
         }?;
         let input_desc = real_executor.identity().to_string();
