+++ src/batch/src/executor2/generate_series.rs
@@ -12,27 +12,111 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::ops::AddAssign;
+use std::sync::Arc;
+
 use futures_async_stream::try_stream;
-use risingwave_common::array::{Array, DataChunk};
+use risingwave_common::array::column::Column;
+use risingwave_common::array::{
+    Array, ArrayBuilder, DataChunk, I32Array, IntervalArray, NaiveDateTimeArray,
+};
 use risingwave_common::catalog::{Field, Schema};
 use risingwave_common::error::{ErrorCode, Result, RwError};
-use risingwave_common::types::DataType;
+use risingwave_common::types::{DataType, Scalar};
+use risingwave_common::util::chunk_coalesce::DEFAULT_CHUNK_BUFFER_SIZE;
 use risingwave_expr::expr::build_from_prost;
 use risingwave_pb::batch_plan::plan_node::NodeBody;
 
+use super::{BoxedExecutor2, BoxedExecutor2Builder};
 use crate::executor::ExecutorBuilder;
-use crate::executor2::{
-    BoxedDataChunkStream, BoxedExecutor2, BoxedExecutor2Builder, Executor2,
-    GenerateSeriesI32Executor2, GenerateSeriesTimestampExecutor2,
-};
+use crate::executor2::{BoxedDataChunkStream, Executor2};
 
-pub struct GenerateSeriesExecutor2Wrapper {
-    input: BoxedExecutor2,
+pub struct GenerateSeriesExecutor2<T: Array, S: Array> {
+    start: T::OwnedItem,
+    stop: T::OwnedItem,
+    step: S::OwnedItem,
+
     schema: Schema,
     identity: String,
 }
 
-impl BoxedExecutor2Builder for GenerateSeriesExecutor2Wrapper {
+impl<T: Array, S: Array> GenerateSeriesExecutor2<T, S> {
+    pub fn new(
+        start: T::OwnedItem,
+        stop: T::OwnedItem,
+        step: S::OwnedItem,
+        schema: Schema,
+        identity: String,
+    ) -> Self {
+        Self {
+            start,
+            stop,
+            step,
+            schema,
+            identity,
+        }
+    }
+}
+
+impl<T: Array, S: Array> Executor2 for GenerateSeriesExecutor2<T, S>
+where
+    T::OwnedItem: PartialOrd<T::OwnedItem>,
+    T::OwnedItem: for<'a> AddAssign<S::RefItem<'a>>,
+{
+    fn schema(&self) -> &Schema {
+        &self.schema
+    }
+
+    fn identity(&self) -> &str {
+        &self.identity
+    }
+
+    fn execute(self: Box<Self>) -> BoxedDataChunkStream {
+        self.do_execute()
+    }
+}
+
+impl<T, S> GenerateSeriesExecutor2<T, S>
+where
+    T: Array,
+    S: Array,
+    T::OwnedItem: PartialOrd<T::OwnedItem>,
+    T::OwnedItem: for<'a> AddAssign<S::RefItem<'a>>,
+{
+    #[try_stream(boxed, ok = DataChunk, error = RwError)]
+    async fn do_execute(self: Box<Self>) {
+        let Self {
+            start, stop, step, ..
+        } = *self;
+
+        // let mut rest_rows = ((stop - start) / step + 1) as usize;
+        let mut cur = start;
+
+        // Simulate a do-while loop.
+        while cur <= stop {
+            let chunk_size = DEFAULT_CHUNK_BUFFER_SIZE;
+            let mut builder = T::Builder::new(chunk_size)?;
+
+            for _ in 0..chunk_size {
+                if cur > stop {
+                    break;
+                }
+                builder.append(Some(cur.as_scalar_ref())).unwrap();
+                cur += step.as_scalar_ref();
+            }
+
+            let arr = builder.finish()?;
+            let columns = vec![Column::new(Arc::new(arr.into()))];
+            let chunk: DataChunk = DataChunk::builder().columns(columns).build();
+
+            yield chunk;
+        }
+    }
+}
+
+pub struct GenerateSeriesExecutor2Builder {}
+
+impl BoxedExecutor2Builder for GenerateSeriesExecutor2Builder {
     fn new_boxed_executor2(source: &ExecutorBuilder) -> Result<BoxedExecutor2> {
         let node = try_match_expand!(
             source.plan_node().get_node_body().unwrap(),
@@ -59,22 +143,15 @@ impl BoxedExecutor2Builder for GenerateSeriesExecutor2Wrapper {
                 if let (Some(start), Some(stop), Some(step)) = (start, stop, step) {
                     let schema = Schema::new(vec![Field::unnamed(DataType::Timestamp)]);
 
-                    let input = Box::new(GenerateSeriesTimestampExecutor2::new(
-                        start,
-                        stop,
-                        step,
-                        schema.clone(),
-                        identity.clone(),
-                    ));
-
-                    Ok(Box::new(Self {
-                        input,
-                        schema,
-                        identity,
-                    }))
+                    Ok(Box::new(GenerateSeriesExecutor2::<
+                        NaiveDateTimeArray,
+                        IntervalArray,
+                    >::new(
+                        start, stop, step, schema, identity
+                    )))
                 } else {
                     Err(ErrorCode::InternalError(
-                        "the parameters of Generate Series Function are incorrect".to_string(),
+                        "the parameters of Generate SeriesFunction are incorrect".to_string(),
                     )
                     .into())
                 }
@@ -87,19 +164,11 @@ impl BoxedExecutor2Builder for GenerateSeriesExecutor2Wrapper {
                 if let (Some(start), Some(stop), Some(step)) = (start, stop, step) {
                     let schema = Schema::new(vec![Field::unnamed(DataType::Int32)]);
 
-                    let input = Box::new(GenerateSeriesI32Executor2::new(
-                        start,
-                        stop,
-                        step,
-                        schema.clone(),
-                        identity.clone(),
-                    ));
-
-                    Ok(Box::new(Self {
-                        input,
-                        schema,
-                        identity,
-                    }))
+                    Ok(Box::new(
+                        GenerateSeriesExecutor2::<I32Array, I32Array>::new(
+                            start, stop, step, schema, identity,
+                        ),
+                    ))
                 } else {
                     Err(ErrorCode::InternalError(
                         "the parameters of Generate Series Function are incorrect".to_string(),
@@ -115,28 +184,91 @@ impl BoxedExecutor2Builder for GenerateSeriesExecutor2Wrapper {
     }
 }
 
-impl Executor2 for GenerateSeriesExecutor2Wrapper {
-    fn schema(&self) -> &Schema {
-        &self.schema
+#[cfg(test)]
+mod tests {
+    use futures::StreamExt;
+    use risingwave_common::array::{Array, ArrayImpl, I32Array};
+    use risingwave_common::catalog::{Field, Schema};
+    use risingwave_common::try_match_expand;
+    use risingwave_common::types::{DataType, IntervalUnit, NaiveDateTimeWrapper};
+    use risingwave_expr::vector_op::cast::str_to_timestamp;
+
+    use super::*;
+
+    #[tokio::test]
+    async fn test_generate_i32_series() {
+        generate_series_test_case(2, 4, 1).await;
+        generate_series_test_case(0, 9, 2).await;
+        generate_series_test_case(0, (DEFAULT_CHUNK_BUFFER_SIZE * 2 + 3) as i32, 1).await;
     }
 
-    fn identity(&self) -> &str {
-        &self.identity
-    }
+    async fn generate_series_test_case(start: i32, stop: i32, step: i32) {
+        let executor = Box::new(GenerateSeriesExecutor2::<I32Array, I32Array> {
+            start,
+            stop,
+            step,
+            schema: Schema::new(vec![Field::unnamed(DataType::Int32)]),
+            identity: "GenerateSeriesExecutor2".to_string(),
+        });
+        let mut remained_values = ((stop - start) / step + 1) as usize;
+        let mut stream = executor.execute();
+        while remained_values > 0 {
+            let chunk = stream.next().await.unwrap().unwrap();
+            let col = chunk.column_at(0);
+            let arr = try_match_expand!(col.array_ref(), ArrayImpl::Int32).unwrap();
 
-    fn execute(self: Box<Self>) -> BoxedDataChunkStream {
-        self.do_execute()
-    }
-}
-
-impl GenerateSeriesExecutor2Wrapper {
-    #[try_stream(boxed, ok = DataChunk, error = RwError)]
-    async fn do_execute(self: Box<Self>) {
-        let input = self.input.execute();
-        #[for_await]
-        for data_chunk in input {
-            let data_chunk = data_chunk?;
-            yield data_chunk
+            if remained_values > DEFAULT_CHUNK_BUFFER_SIZE {
+                assert_eq!(arr.len(), DEFAULT_CHUNK_BUFFER_SIZE);
+            } else {
+                assert_eq!(arr.len(), remained_values);
+            }
+            remained_values -= arr.len();
         }
+        assert!(stream.next().await.is_none());
+    }
+
+    #[tokio::test]
+    async fn test_generate_time_series() {
+        let start_time = str_to_timestamp("2008-03-01 00:00:00").unwrap();
+        let stop_time = str_to_timestamp("2008-03-09 00:00:00").unwrap();
+        let one_minute_step = IntervalUnit::from_minutes(1);
+        let one_hour_step = IntervalUnit::from_minutes(60);
+        let one_day_step = IntervalUnit::from_days(1);
+        generate_time_series_test_case(start_time, stop_time, one_minute_step).await;
+        generate_time_series_test_case(start_time, stop_time, one_hour_step).await;
+        generate_time_series_test_case(start_time, stop_time, one_day_step).await;
+    }
+
+    async fn generate_time_series_test_case(
+        start: NaiveDateTimeWrapper,
+        stop: NaiveDateTimeWrapper,
+        step: IntervalUnit,
+    ) {
+        let executor = Box::new(
+            GenerateSeriesExecutor2::<NaiveDateTimeArray, IntervalArray>::new(
+                start,
+                stop,
+                step,
+                Schema::new(vec![Field::unnamed(DataType::Int32)]),
+                "GenerateSeriesTimestampExecutor2".to_string(),
+            ),
+        );
+        let interval_ms = step.get_total_ms();
+        let start_stop_spread_ms = stop.sub(start).num_milliseconds();
+        let mut remained_values = (start_stop_spread_ms / interval_ms + 1) as usize;
+        let mut stream = executor.execute();
+        while remained_values > 0 {
+            let chunk = stream.next().await.unwrap().unwrap();
+            let col = chunk.column_at(0);
+            let arr = try_match_expand!(col.array_ref(), ArrayImpl::NaiveDateTime).unwrap();
+
+            if remained_values > DEFAULT_CHUNK_BUFFER_SIZE {
+                assert_eq!(arr.len(), DEFAULT_CHUNK_BUFFER_SIZE);
+            } else {
+                assert_eq!(arr.len(), remained_values);
+            }
+            remained_values -= arr.len();
+        }
+        assert!(stream.next().await.is_none());
     }
 }
