+++ src/common/src/types/chrono_wrapper.rs
@@ -15,9 +15,11 @@
 use std::fmt::{Display, Formatter};
 use std::hash::{Hash, Hasher};
 use std::io::Write;
+use std::ops::AddAssign;
 
 use chrono::{Datelike, Duration, NaiveDate, NaiveDateTime, NaiveTime, Timelike};
 
+use super::IntervalUnit;
 use crate::error::ErrorCode::{InternalError, IoError};
 use crate::error::{Result, RwError};
 use crate::util::value_encoding::error::ValueEncodingError;
@@ -192,6 +194,13 @@ impl NaiveDateTimeWrapper {
 
     pub fn add(&self, duration: Duration) -> Self {
         NaiveDateTimeWrapper::new(self.0 + duration)
+    }
+}
+
+impl AddAssign<IntervalUnit> for NaiveDateTimeWrapper {
+    fn add_assign(&mut self, rhs: IntervalUnit) {
+        let durantion: Duration = rhs.into();
+        *self = NaiveDateTimeWrapper::new(self.0 + durantion);
     }
 }
